[migrate] modify migrations according to ERD
