+++ laravel/database/seeders/DatabaseSeeder.php
@@ -5,8 +5,7 @@ namespace Database\Seeders;
 use Database\Seeders\AssignmentSeeder;
 use Database\Seeders\CommentSeeder;
 use Database\Seeders\CourseSeeder;
-use Database\Seeders\HomeworkSeeder;
-use Database\Seeders\RoleSeeder;
+use Database\Seeders\StudentAssignmentSeeder;
 use Database\Seeders\UserSeeder;
 use Illuminate\Database\Eloquent\Model; 
 use Illuminate\Database\Seeder;
@@ -25,7 +24,7 @@ class DatabaseSeeder extends Seeder
         $this->call(UserSeeder::class);
         $this->call(CourseSeeder::class);
         $this->call(AssignmentSeeder::class);
-        $this->call(HomeworkSeeder::class);
+        $this->call(StudentAssignmentSeeder::class);
         $this->call(CommentSeeder::class);
         $this->call(StudentCourseSeeder::class);
         $this->call(TeacherCourseSeeder::class);
